# control-plane/rest/openapi-specs/v0_api_spec.yaml
openapi: 3.0.3
info:
  title: Mayastor RESTful API
  version: v0
servers:
  - url: /v0
paths:
...
  "/nexuses/{nexus_id}/children":
    get:
      tags:
        - Children
      operationId: get_nexus_children
      parameters:
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Child"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nexuses/{nexus_id}/children/{child_id:.*}":
    get:
      tags:
        - Children
      operationId: get_nexus_child
      x-actix-query-string: true
      parameters:
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: "child_id:.*"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    put:
      tags:
        - Children
      operationId: put_nexus_child
      x-actix-query-string: true
      parameters:
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: "child_id:.*"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    delete:
      tags:
        - Children
      operationId: del_nexus_child
      x-actix-query-string: true
      parameters:
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: "child_id:.*"
          required: true
          schema:
            type: string
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
  "/nodes/{node_id}/nexuses/{nexus_id}/children":
    get:
      tags:
        - Children
      operationId: get_node_nexus_children
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Child"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
  "/nodes/{node_id}/nexuses/{nexus_id}/children/{child_id:.*}":
    get:
      tags:
        - Children
      operationId: get_node_nexus_child
      x-actix-query-string: true
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: "child_id:.*"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    put:
      tags:
        - Children
      operationId: put_node_nexus_child
      x-actix-query-string: true
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: "child_id:.*"
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
    delete:
      tags:
        - Children
      operationId: del_node_nexus_child
      x-actix-query-string: true
      parameters:
        - in: path
          name: node_id
          required: true
          schema:
            type: string
        - in: path
          name: nexus_id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: "child_id:.*"
          required: true
          schema:
            type: string
      responses:
        "204":
          description: OK
        "400":
          description: Request Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "408":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "412":
          description: Precondition Failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "416":
          description: Range Not satisfiable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "501":
          description: Not Implemented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "504":
          description: Gateway Timeout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
        "507":
          description: Insufficient Storage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RestJsonError"
      security:
        - JWT: []
...
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ChildState:
      example: Online
      description: State of a Nexus Child
      type: string
      enum:
        - Unknown
        - Online
        - Degraded
        - Faulted
    Child:
      example:
        rebuildProgress: ~
        state: Online
        uri: "nvmf://10.1.0.5:8420/nqn.2019-05.io.openebs:37d83441-e8ef-4e17-a29e-25169d91cb96"
      description: Child information
      type: object
      properties:
        rebuildProgress:
          description: current rebuild progress (%)
          type: integer
          minimum: 0
        state:
          description: state of the child
          allOf:
            - $ref: "#/components/schemas/ChildState"
        uri:
          description: uri of the child device
          type: string
      required:
        - state
        - uri
...
    RestJsonError:
      example:
        details: "The Pool 'pooloop' was not found"
        kind: NotFound
      description: Rest Json Error format
      type: object
      properties:
        details:
          description: detailed error information
          type: string
        kind:
          description: error kind
          type: string
          enum:
            - Timeout
            - Deserialize
            - Internal
            - InvalidArgument
            - DeadlineExceeded
            - NotFound
            - AlreadyExists
            - PermissionDenied
            - ResourceExhausted
            - FailedPrecondition
            - NotShared
            - NotPublished
            - AlreadyPublished
            - AlreadyShared
            - Aborted
            - OutOfRange
            - Unimplemented
            - Unavailable
            - Unauthenticated
            - Unauthorized
            - Conflict
            - FailedPersist
            - Deleting
      required:
        - details
        - kind
...