# openapi/src/apis/children_api.rs
#![allow(
    missing_docs,
    trivial_casts,
    unused_variables,
    unused_mut,
    unused_imports,
    unused_extern_crates,
    non_camel_case_types
)]

use actix_web::web::{self, Json, Path, Query};

#[async_trait::async_trait]
pub trait ChildrenApi {
    async fn del_nexus_child(
        query: &str,
        Path((nexus_id, child_id_)): Path<(String, String)>,
    ) -> Result<(), crate::apis::RestError<crate::models::RestJsonError>>;
    async fn del_node_nexus_child(
        query: &str,
        Path((node_id, nexus_id, child_id_)): Path<(String, String, String)>,
    ) -> Result<(), crate::apis::RestError<crate::models::RestJsonError>>;
    async fn get_nexus_child(
        query: &str,
        Path((nexus_id, child_id_)): Path<(String, String)>,
    ) -> Result<Json<crate::models::Child>, crate::apis::RestError<crate::models::RestJsonError>>;
    async fn get_nexus_children(
        Path(nexus_id): Path<String>,
    ) -> Result<Json<Vec<crate::models::Child>>, crate::apis::RestError<crate::models::RestJsonError>>;
    async fn get_node_nexus_child(
        query: &str,
        Path((node_id, nexus_id, child_id_)): Path<(String, String, String)>,
    ) -> Result<Json<crate::models::Child>, crate::apis::RestError<crate::models::RestJsonError>>;
    async fn get_node_nexus_children(
        Path((node_id, nexus_id)): Path<(String, String)>,
    ) -> Result<Json<Vec<crate::models::Child>>, crate::apis::RestError<crate::models::RestJsonError>>;
    async fn put_nexus_child(
        query: &str,
        Path((nexus_id, child_id_)): Path<(String, String)>,
    ) -> Result<Json<crate::models::Child>, crate::apis::RestError<crate::models::RestJsonError>>;
    async fn put_node_nexus_child(
        query: &str,
        Path((node_id, nexus_id, child_id_)): Path<(String, String, String)>,
    ) -> Result<Json<crate::models::Child>, crate::apis::RestError<crate::models::RestJsonError>>;
}